requirements.txt
pandas
numpy
seaborn
matplotlib.pyplot